Attempted to add isStarred method
Didn't end up working.  Could get a return value inside jni from the
task looper.  Realized this is a known issue listed from PsyOnSpotify.
Left code in place just in case.  Tried this so we could have the star
in the player be accurate for more than just the session.