Use the intro sound as an awesome sound check mechanism
On (at least) Samsung Note 5, tunnel.decode property returns empty but
analysis through the java Visualizer class fails.  So use the intro
sound to see if we'll get any data in that class.

TunnelPlayer Workaround not needed, which apparently we weren't even
using (it has to be initialized), and even when used apparently doesn't
work on all devices.

The good reference I've been using:
https://github.com/felixpalmer/android-visualizer/issues/5